Move release and release request builders to project
